WIP: Script Editor: use the framework to choose files
This calls the UIService's chooseFile method, rather than hardcoding
usage of JFileChooser. This is especially important for scenarios like
the LegacyUI being visible, since in many ways it reuses the SwingUI,
but with many functions implemented differently, including chooseFile.

The reason this is a "WIP" is because it does not preserve the file
filtering feature of the previous implementation. In particular, it does
not limit selection of files to only JAR files when exporting a JAR.

Before this branch can be completed, we will need to enhance the
UIService and UserInterface APIs to be powerful enough to do that in an
agnostic way.